Check trace at TE spec test in ModelCheckerTestCase.
This is one approach of doing things which test ALL the TE spec traces
with the original ones. We are able to do more powerful checks with
each EC.* code.

The tests failing are the ones which contains a `ALIAS` and `VIEW`,
should we add support for these in our TE spec generation?